pyproject.toml
[tool.poetry]
name = "django-todo"
version = "0.1.0"
description = ""
authors = ["rkjar <rkjar@yandex.ru>"]
readme = "README.md"
packages = [{include = "django_todo"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^3.2"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
drf-spectacular = "^0.26.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
